Revert "PR #64 - major refactor of copy/paste connect() methods out of service classes"
This reverts commit 47a060275a1f2bb8a1929545967dbb0607437918.

This refactor didn't work out. When we set
self.connect_function = boto.something.some_function
it becomes a bound method of ``self``, apparently. The end result
is that ``self`` ALWAYS gets passed in as the first argument, and there's
really no way around this. The only alternative would be to set connect_function
to a string, the name of the function to call, but then testing becomes
horrible and it also won't support functions *or* classes.