s3: Only find buckets in our region
Quotas for S3 buckets are regional. Currently though, we are returning
buckets in all regions when we query the API. This means that we sum all
buckets in all regions against the quota for each region (modulo issue

We can fix this by doing a `head_bucket` call for each bucket, and only
counting the ones that are in our current region.

Note that this needs the `s3:ListBucket` IAM permission.